replace WithAvatarBadge and WithNameBadge with DisplayBadge
